chore(#106): refactor oversized files to meet 500-line limit
